backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
PyMongo==4.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
google-generativeai==0.3.2
requests==2.31.0
pytest==7.4.2
pytest-flask==1.2.0

